# Conflicts:
#	pom.xml
#	src/main/java/com/windear/app/controller/BookController.java
#	src/main/java/com/windear/app/controller/NewsController.java
#	src/main/java/com/windear/app/entity/Book.java
#	src/main/java/com/windear/app/entity/News.java
#	src/main/java/com/windear/app/exception/ErrorResponse.java
#	src/main/java/com/windear/app/exception/GlobalExceptionHandler.java
#	src/main/java/com/windear/app/service/NewsService.java
#	src/main/java/com/windear/app/service/NewsServiceImpl.java
#	src/main/resources/application.properties

+++ src/main/java/com/windear/app/service/NewsServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -27,18 +28,40 @@ public class NewsServiceImpl implements NewsService{
     @Override
     public News findNewsById(int newsId) {
         Optional<News> news = newsRepository.findById(newsId);
-        if(news.isPresent()) {
+        if (news.isPresent()) {
             return news.get();
-        }
-        else {
+        } else {
             throw new RuntimeException("User with title not found: " + newsId);
         }
     }
 
     @Override
+    public List<News> findAllNews() {
+        return newsRepository.findAll();
+    }
+
+    @Override
+    @Transactional
+    public News update(News news) {
+        News newsFromDB = findNewsById(news.getNewsId());
+        if(news.getDescription() != null) {
+            newsFromDB.setDescription(news.getDescription());
+        }
+        if(news.getTitle() != null) {
+            newsFromDB.setTitle(news.getTitle());
+        }
+        if(news.getImageUrl() != null) {
+            newsFromDB.setImageUrl(news.getImageUrl());
+        }
+        return newsRepository.save(newsFromDB);
+    }
+
+    @Override
     @Transactional
     public void deleteNews(int newsId) {
-        newsRepository.deleteById(newsId);
+        if(findNewsById(newsId) != null) {
+            newsRepository.deleteById(newsId);
+        }
     }
 
 }
